requirements.txt
kafka-python>=2.3.0,<3.0
mmh3>=4.0.0
boto3>=1.34.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Linting & Type Checking
ruff>=0.1.0
mypy>=1.7.0
types-requests>=2.31.0
